fix: low-severity / hardening items from the full source audit
Final pass over the audit's low-severity findings. Touched-area suites (3196 tests)
pass with no regressions.

- IntervalsIndexedFormat.cpp: the 2D streaming index finalizer seeked past a 36-byte
  (1D) header and wrote 24-byte entries, but IntervalsIndex2D::load expects a 40-byte
  header and 28-byte entries (with a trailing reserved field) - every entry was
  misaligned and the set was unreadable. Now seeks sizeof(Index2DHeader) and writes
  the 28-byte entry. Gated behind the non-default gmulticontig.indexed_format=TRUE,
  but data loss when enabled.
- BufferedFile::getc(): return via unsigned char so a byte >= 0x80 is not sign-extended
  to a negative int that looks like EOF - this truncated non-ASCII track attribute
  values on read.
- GenomeEditImplant.cpp (ggenome.implant): used Rf_error (longjmp) inside its
  try/catch, bypassing the C++ destructors and leaking the 256MB sequence buffer and
  I/O buffers on error paths. Added RdbInitializer + verror so the catch runs and
  destructors clean up (matches every other misha C entry point).
- StatQuadTreeCached.h get_chunk(): reject a corrupt on-disk chunk size < 8 before
  allocating (size < 8 overflowed the *(int64_t*) store; size - sizeof(size)
  underflowed the read length).
- created.by provenance: gtrack.2d.create deparsed `intervals` after reassigning it
  (emitting a structure() dump); gtrack.2d.import_contacts wrote a quoted "NULL" for a
  NULL fends; gtrack.modify prefixed "NA\n" when the prior attribute was absent.
- track-array.R .gslice: reject NA slice indices instead of forwarding NA to C.
- zzz.R: floor gmax.processes at 1 (was 0 on single-core / undetectable-core hosts).
- IntervalValidator.cpp restrict_bins: ceil() now over real division (was a no-op over
  integer division; currently dead code).

Not fixed (with reason): rdbutils/TrackExpressionVars PROTECT bookkeeping (benign -
objects stay reachable via the GVTRACKS env binding; runprotect stack reclaimed by
RdbInitializer); neighbors.R directional dist==0 double-count (a semantic limitation -
dist==0 is the same value on either side, so value filtering cannot assign a
direction); DnaPSSM N-base scoring inconsistency (numeric change, no clear correct
path, risks altering PWM results - needs benchmarking).

Regression tests: test-audit-2026-06-28-lows.R (L5/L7/L12).

+++ src/IntervalValidator.cpp
@@ -15,7 +15,9 @@ IntervalValidator::IntervalValidator(const GenomeChromKey &chromkey) :
 void IntervalValidator::restrict_bins(int64_t maxbins, GIntervals &intervals, unsigned binsize) const
 {
 	for (GIntervals::const_iterator iinterval = intervals.begin(); iinterval != intervals.end(); ++iinterval) {
-		int64_t bins = max((int64_t)0, (int64_t)ceil(iinterval->end / binsize) - (int64_t)(iinterval->start / binsize));
+		// cast to double so ceil() rounds the real division up (end/binsize is
+		// integer division otherwise, making ceil a no-op and undercounting bins).
+		int64_t bins = max((int64_t)0, (int64_t)ceil((double)iinterval->end / binsize) - (int64_t)(iinterval->start / binsize));
 
 		if (bins > maxbins)
 			verror("The interval %s [%ld, %ld) covers too wide range of samples that might cause memory allocation failure.\n"
